ShowData 1.0/requirements.txt
requests
pandas
sqlalchemy
matplotlib
seaborn
python-dotenv
